Renamed and reworked the native disector to allow us to have two layers of disection
