+++ etc/dissectors/n5305a/dissectors.h
@@ -4,4 +4,7 @@
 extern void registerProtocolN5305AFraming(void);
 extern void registerDissectorN5305AFraming(void);
 
+extern void registerProtocolN5305ATransaction(void);
+extern void registerDissectorN5305ATransaction(void);
+
 #endif /*N5305A_FRAME_DISSECTOR__H*/
